test: continue scanner coverage decomposition
